revamp layout for new view, move MapContainer state/search to AddOpportunity
